+++ src/locales/fi/index.js
@@ -2,12 +2,8 @@
 // Licensed under the MIT License
 
 export default {
-  indexTitle: 'Etusivu',
-  indexServicesId: 'palvelut',
-  indexServicesTitle: 'Mitä linssimme taittaa?',
-  indexReadMore: 'Lue lisää',
-  indexPortfolioId: 'portfolio',
-  indexPortfolioTitle: 'Portfolio',
-  indexContactId: 'yhteystiedot',
-  indexContactTitle: 'Yhteystiedot',
+  indexContactFormName: 'Nimi',
+  indexContactFormEmail: 'Sähköposti',
+  indexContactFormTel: 'Puhelinnumero',
+  indexContactFormMessage: 'Viestisi meille',
 };