+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [0.15.0] - 2021-03-27
+
+### Added
+
+- Link to the terms of use to the footer.
+- Contact for to the front page.
+
+### Changed
+
+- Button colour to be set by using an explicit colour option.
+- Colour of the green buttons to be a bit darker.
+- Placeholder for some of the images to be blurred instead of single colour.
+- Search box to always have a shadow.
+- Box shadows to show better in dark mode.
+
 ## [0.14.2] - 2021-03-27
 
 ### Fixed
@@ -380,7 +395,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - Support for internationalizing pages.
 - Both light and dark themes for the site.
 
-[unreleased]: https://github.com/visiosto/visiosto.fi/compare/v0.14.2...HEAD
+[unreleased]: https://github.com/visiosto/visiosto.fi/compare/v0.15.0...HEAD
+[0.15.0]: https://github.com/visiosto/visiosto.fi/compare/v0.14.2...v0.15.0
 [0.14.2]: https://github.com/visiosto/visiosto.fi/compare/v0.14.1...v0.14.2
 [0.14.1]: https://github.com/visiosto/visiosto.fi/compare/v0.14.0...v0.14.1
 [0.14.0]: https://github.com/visiosto/visiosto.fi/compare/v0.13.0...v0.14.0